Add files via upload
This dataset has 14941 rows and 2 columns. This project is uses Natural Language Processing technique in order to analyze the tweets data which is extracted from Twitter based on demonetization. We have analyzed the tweets based on the sentiments(positive, negative, neutral), checked out the mentions of people tweeting, checked the mentions of PM to understand the opinions of people on PM regarding this drastic step. Along with that hashtag analysis was done followed by preposition analysis of the tweets and all these steps were done after the text cleaning and pre-processing.